requirements.txt
selenium==4.1.0
pytest==7.0.1
pyyaml==6.0
pytest-xdist==2.5.0